added listener

+++ src/main/java/org/qcri/micromappers/utility/HttpDownloadUtility.java
@@ -13,7 +13,7 @@ import java.net.URL;
 public class HttpDownloadUtility {
     private static final int BUFFER_SIZE = 4096;
 
-    public boolean downloadFile(String fileURL, String saveDir){
+    public boolean downloadFile(String fileURL, String saveDir, String fileName){
         boolean downloadFileCompleted = false;
         HttpURLConnection httpConn = null;
         try{
@@ -23,23 +23,12 @@ public class HttpDownloadUtility {
 
             // always check HTTP response code first
             if (responseCode == HttpURLConnection.HTTP_OK) {
-                String fileName = "";
+
                 String disposition = httpConn.getHeaderField("Content-Disposition");
                 String contentType = httpConn.getContentType();
                 int contentLength = httpConn.getContentLength();
 
-                if (disposition != null) {
-                    // extracts file name from header field
-                    int index = disposition.indexOf("filename=");
-                    if (index > 0) {
-                        fileName = disposition.substring(index + 10,
-                                disposition.length() - 1);
-                    }
-                } else {
-                    // extracts file name from URL
-                    fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1,
-                            fileURL.length());
-                }
+                fileName = getFileName(fileURL, disposition, fileName);
 
                 System.out.println("Content-Type = " + contentType);
                 System.out.println("Content-Disposition = " + disposition);
@@ -49,8 +38,9 @@ public class HttpDownloadUtility {
                 // opens input stream from the HTTP connection
                 InputStream inputStream = httpConn.getInputStream();
                 String saveFilePath = saveDir + File.separator + fileName;
+                String saveFilePathProcessed = saveDir + File.separator + fileName + FilePathSpec.GDELT_PROCESSED_EXTENTION;
                 // check file exists
-                if(! new File(saveFilePath).exists()){
+                if(!new File(saveFilePath).exists() && !new File(saveFilePathProcessed).exists()){
                     // opens an output stream to save into file
                     FileOutputStream outputStream = new FileOutputStream(saveFilePath);
 
@@ -83,5 +73,27 @@ public class HttpDownloadUtility {
         }
 
         return downloadFileCompleted;
+    }
+
+    private String getFileName(String fileURL, String disposition, String fileName){
+
+        if(fileName!=null && !fileName.isEmpty()){
+            return fileName;
+        }
+
+        if (disposition != null) {
+            // extracts file name from header field
+            int index = disposition.indexOf("filename=");
+            if (index > 0) {
+                fileName = disposition.substring(index + 10,
+                        disposition.length() - 1);
+            }
+        } else {
+            // extracts file name from URL
+            fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1,
+                    fileURL.length());
+        }
+
+        return fileName;
     }
 }
